Move code for reading process flows into input/process/flow.rs
