+++ src/input/process.rs
@@ -2,21 +2,22 @@
 use crate::commodity::Commodity;
 use crate::input::region::{define_region_id_getter, read_regions_for_entity};
 use crate::input::*;
-use crate::process::{FlowType, Process, ProcessFlow, ProcessParameter};
+use crate::process::{Process, ProcessFlow, ProcessParameter};
 use crate::time_slice::TimeSliceInfo;
 use ::log::warn;
 use anyhow::{ensure, Context, Result};
 use itertools::Itertools;
-use serde::{Deserialize, Deserializer};
+use serde::Deserialize;
 use std::collections::{HashMap, HashSet};
 use std::ops::RangeInclusive;
 use std::path::Path;
 use std::rc::Rc;
 pub mod availability;
 use availability::read_process_availabilities;
+pub mod flow;
+use flow::read_process_flows;
 
 const PROCESSES_FILE_NAME: &str = "processes.csv";
-const PROCESS_FLOWS_FILE_NAME: &str = "process_flows.csv";
 const PROCESS_PACS_FILE_NAME: &str = "process_pacs.csv";
 const PROCESS_PARAMETERS_FILE_NAME: &str = "process_parameters.csv";
 const PROCESS_REGIONS_FILE_NAME: &str = "process_regions.csv";
@@ -31,32 +32,6 @@ macro_rules! define_process_id_getter {
     };
 }
 use define_process_id_getter;
-
-define_process_id_getter! {ProcessFlow}
-
-#[derive(PartialEq, Debug, Deserialize)]
-struct ProcessFlowRaw {
-    process_id: String,
-    commodity_id: String,
-    flow: f64,
-    #[serde(default)]
-    flow_type: FlowType,
-    #[serde(deserialize_with = "deserialise_flow_cost")]
-    flow_cost: f64,
-}
-define_process_id_getter! {ProcessFlowRaw}
-
-/// Custom deserialiser for flow cost - treat empty fields as 0.0
-fn deserialise_flow_cost<'de, D>(deserialiser: D) -> Result<f64, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    let value: Option<f64> = Deserialize::deserialize(deserialiser)?;
-    match value {
-        None => Ok(0.0),
-        Some(value) => Ok(value),
-    }
-}
 
 /// Primary Activity Commodity
 #[derive(PartialEq, Clone, Eq, Hash, Debug, Deserialize)]
@@ -173,42 +148,6 @@ struct ProcessDescription {
     description: String,
 }
 define_id_getter! {ProcessDescription}
-
-/// Read 'ProcessFlowRaw' records from an iterator and convert them into 'ProcessFlow' records.
-fn read_process_flows_from_iter<I>(
-    iter: I,
-    process_ids: &HashSet<Rc<str>>,
-    commodities: &HashMap<Rc<str>, Rc<Commodity>>,
-) -> Result<HashMap<Rc<str>, Vec<ProcessFlow>>>
-where
-    I: Iterator<Item = ProcessFlowRaw>,
-{
-    iter.map(|flow_raw| -> Result<ProcessFlow> {
-        let commodity = commodities
-            .get(flow_raw.commodity_id.as_str())
-            .with_context(|| format!("{} is not a valid commodity ID", &flow_raw.commodity_id))?;
-
-        Ok(ProcessFlow {
-            process_id: flow_raw.process_id,
-            commodity: Rc::clone(commodity),
-            flow: flow_raw.flow,
-            flow_type: flow_raw.flow_type,
-            flow_cost: flow_raw.flow_cost,
-        })
-    })
-    .process_results(|iter| iter.into_id_map(process_ids))?
-}
-
-fn read_process_flows(
-    model_dir: &Path,
-    process_ids: &HashSet<Rc<str>>,
-    commodities: &HashMap<Rc<str>, Rc<Commodity>>,
-) -> Result<HashMap<Rc<str>, Vec<ProcessFlow>>> {
-    let file_path = model_dir.join(PROCESS_FLOWS_FILE_NAME);
-    let process_flow_csv = read_csv(&file_path)?;
-    read_process_flows_from_iter(process_flow_csv, process_ids, commodities)
-        .with_context(|| input_err_msg(&file_path))
-}
 
 fn read_process_parameters_from_iter<I>(
     iter: I,
@@ -413,11 +352,10 @@ pub fn read_processes(
 
 #[cfg(test)]
 mod tests {
-
-    use crate::commodity::{CommodityCostMap, CommodityType};
-    use crate::time_slice::TimeSliceLevel;
-
     use super::*;
+    use crate::commodity::{CommodityCostMap, CommodityType};
+    use crate::process::FlowType;
+    use crate::time_slice::TimeSliceLevel;
 
     fn create_param_raw(
         start_year: Option<u32>,
@@ -554,129 +492,6 @@ mod tests {
         assert!(
             create_param_raw(Some(2000), Some(2100), 0, Some(1.0), Some(-1.0))
                 .validate()
-                .is_err()
-        );
-    }
-
-    #[test]
-    fn test_read_process_flows_from_iter_good() {
-        let process_ids = ["id1".into(), "id2".into()].into_iter().collect();
-        let commodities: HashMap<Rc<str>, Rc<Commodity>> = ["commodity1", "commodity2"]
-            .into_iter()
-            .map(|id| {
-                let commodity = Commodity {
-                    id: id.into(),
-                    description: "Some description".into(),
-                    kind: CommodityType::InputCommodity,
-                    time_slice_level: TimeSliceLevel::Annual,
-                    costs: CommodityCostMap::new(),
-                    demand_by_region: HashMap::new(),
-                };
-
-                (Rc::clone(&commodity.id), commodity.into())
-            })
-            .collect();
-
-        let flows_raw = [
-            ProcessFlowRaw {
-                process_id: "id1".into(),
-                commodity_id: "commodity1".into(),
-                flow: 1.0,
-                flow_type: FlowType::Fixed,
-                flow_cost: 1.0,
-            },
-            ProcessFlowRaw {
-                process_id: "id1".into(),
-                commodity_id: "commodity2".into(),
-                flow: 1.0,
-                flow_type: FlowType::Fixed,
-                flow_cost: 1.0,
-            },
-            ProcessFlowRaw {
-                process_id: "id2".into(),
-                commodity_id: "commodity1".into(),
-                flow: 1.0,
-                flow_type: FlowType::Fixed,
-                flow_cost: 1.0,
-            },
-        ];
-
-        let expected = HashMap::from([
-            (
-                "id1".into(),
-                vec![
-                    ProcessFlow {
-                        process_id: "id1".into(),
-                        commodity: commodities.get("commodity1").unwrap().clone(),
-                        flow: 1.0,
-                        flow_type: FlowType::Fixed,
-                        flow_cost: 1.0,
-                    },
-                    ProcessFlow {
-                        process_id: "id1".into(),
-                        commodity: commodities.get("commodity2").unwrap().clone(),
-                        flow: 1.0,
-                        flow_type: FlowType::Fixed,
-                        flow_cost: 1.0,
-                    },
-                ],
-            ),
-            (
-                "id2".into(),
-                vec![ProcessFlow {
-                    process_id: "id2".into(),
-                    commodity: commodities.get("commodity1").unwrap().clone(),
-                    flow: 1.0,
-                    flow_type: FlowType::Fixed,
-                    flow_cost: 1.0,
-                }],
-            ),
-        ]);
-
-        let actual =
-            read_process_flows_from_iter(flows_raw.into_iter(), &process_ids, &commodities)
-                .unwrap();
-        assert_eq!(expected, actual);
-    }
-
-    #[test]
-    fn test_read_process_flows_from_iter_bad_commodity_id() {
-        let process_ids = ["id1".into(), "id2".into()].into_iter().collect();
-        let commodities = ["commodity1", "commodity2"]
-            .into_iter()
-            .map(|id| {
-                let commodity = Commodity {
-                    id: id.into(),
-                    description: "Some description".into(),
-                    kind: CommodityType::InputCommodity,
-                    time_slice_level: TimeSliceLevel::Annual,
-                    costs: CommodityCostMap::new(),
-                    demand_by_region: HashMap::new(),
-                };
-
-                (Rc::clone(&commodity.id), commodity.into())
-            })
-            .collect();
-
-        let flows_raw = [
-            ProcessFlowRaw {
-                process_id: "id1".into(),
-                commodity_id: "commodity1".into(),
-                flow: 1.0,
-                flow_type: FlowType::Fixed,
-                flow_cost: 1.0,
-            },
-            ProcessFlowRaw {
-                process_id: "id1".into(),
-                commodity_id: "commodity3".into(),
-                flow: 1.0,
-                flow_type: FlowType::Fixed,
-                flow_cost: 1.0,
-            },
-        ];
-
-        assert!(
-            read_process_flows_from_iter(flows_raw.into_iter(), &process_ids, &commodities)
                 .is_err()
         );
     }
